Moved all remaining test helper code from main helpers file into test helpers
